# openapi.yaml
openapi: 3.0.3
info:
  title: MS Calendar API
  version: 1.0.0
  description: OpenAPI spec split by feature.
servers:
  - url: http://localhost:3201
    description: Local development
  - url: http://localhost:3299/calendar
    description: Local API gateway (prefix /calendar)
...
paths:
  /api/demo/token:
    get:
      tags: [Demo]
      summary: Generate demo JWT token
      description: Development-only endpoint. Returns a ready-to-use JWT so evaluators can authorize protected endpoints directly from Swagger UI.
      parameters:
        - in: query
          name: expiresIn
          required: false
          schema:
            type: string
            example: "2h"
          description: JWT expiration (example 2h, 30m, 1d).
      responses:
        "200":
          description: Demo token generated
          content:
            application/json:
              schema:
                type: object
                required: [ok, token, claims, authorizationHeader]
                properties:
                  ok:
                    type: boolean
                    example: true
                  token:
                    type: string
                  claims:
                    type: object
                    required: [idUser, idCompanySelected, role]
                    properties:
                      idUser:
                        type: string
                        format: uuid
                      idCompanySelected:
                        type: string
                        format: uuid
                      role:
                        type: string
                        example: ROLE_OWNER
                  authorizationHeader:
                    type: string
                    example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "404":
          description: Not available outside development environment
          content:
            application/json:
              schema:
                $ref: "./src/features/event-client/openapi/schemas/index.yaml#/GenericErrorResponse"
        "500":
          description: Missing JWT configuration
          content:
            application/json:
              schema:
                $ref: "./src/features/event-client/openapi/schemas/index.yaml#/GenericErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT